backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
imagekitio==2.2.5
psycopg2-binary==2.8.6
python-dotenv==0.15.0
django-anymail[mailgun]==8.1.0
djangorestframework-camel-case==1.2.0
django-update-from-dict
django-cors-headers==3.5.0